Edit: edit function in supplier

+++ src/supplier.ts
@@ -92,15 +92,20 @@ app.post(
   "/post",
   async ({ body }) => {
     try {
-      const supplierList = await db.$queryRaw`
+      const supplier = await db.$queryRaw`
       INSERT INTO supplier (supplier_name, tax_number)
-      VALUES (${body.supplier_name}, ${body.tax_number});
+      VALUES (
+        ${body.supplier_name}, 
+        ${body.tax_number}
+        )
+      RETURNING supplier_name, tax_number
       `;
-      console.log("Record inserted successfully: ", supplierList);
-      return supplierList;
+
+      console.log("Supplier inserted successfully: ", supplier);
+      return { message: "Supplier inserted successfully", supplier };
     } catch (error) {
-      console.error("Error insrting record: ", error);
-      return { error: "Failed to insert record" };
+      console.error("Error insrting supplier: ", error);
+      return { error: "Failed to insert supplier" };
     }
   },
   {
@@ -118,23 +123,35 @@ app.put(
   "/put",
   async ({ body }) => {
     try {
-      const supplierList: any = await db.$queryRaw`
-    UPDATE supplier
-    SET supplier_name = ${body.supplier_name},
-    tax_number = ${body.tax_number}
-    WHERE supplier_id = ${body.supplier_id}
-    RETURNING supplier_id, supplier_name, tax_number;
-    `;
-      if (supplierList.length > 0) {
-        console.log("Record updated successfully:", supplierList);
-        return supplierList;
+      const updates = [];
+
+      if (body.supplier_name !== undefined) {
+        updates.push(`supplier_name = ${body.supplier_name}`);
+      }
+
+      if (body.tax_number !== undefined) {
+        updates.push(`tax_number = ${body.tax_number}`);
+      }
+
+      const updateFields = updates.join(", ");
+
+      const updatedSupplier: any = await db.$executeRaw`
+      UPDATE supplier
+      SET ${updateFields}
+      WHERE supplier_id = ${body.supplier_id}
+      RETURNING supplier_id, supplier_name, tax_number
+      `;
+
+      if (updatedSupplier.length > 0) {
+        console.log("Supplier updated successfully:", updatedSupplier);
+        return updatedSupplier;
       } else {
         console.log("No record found with the given ID to update");
         return { message: "No record found with the given ID" };
       }
     } catch (error) {
-      console.error("Error updating record: ", error);
-      return { error: "Failed to update record" };
+      console.error("Error updating supplier: ", error);
+      return { error: "Failed to update supplier" };
     }
   },
   {
@@ -153,12 +170,26 @@ app.delete(
   "/delete",
   async ({ body }) => {
     try {
-      const supplierList = await db.$queryRaw`
+      const existingSupplier: any = await db.$queryRaw`
+        SELECT id FROM supplier WHERE supplier_id = ${body.supplier_id};
+      `;
+
+      if (existingSupplier.length === 0) {
+        console.log("No record found with the given ID to delete");
+        return { message: "No record found with the given ID" };
+      }
+
+      const deletedSupplier: any = await db.$queryRaw`
       DELETE FROM supplier
       WHERE supplier_id = ${body.supplier_id}
+      RETURNING supplier_id
       `;
-      console.log("Record deleted successfully");
-      return { message: "Record deleted successfully" };
+
+      console.log("Record deleted successfully: ", deletedSupplier);
+      return {
+        message: "Record deleted successfully",
+        supplier_id: deletedSupplier[0].supplier_id,
+      };
     } catch (error) {
       console.error("Error deleting record: ", error);
       return { error: "Failed to delete record" };
